feat: Lsp4MetadataForm
