+++ src/types.d.ts
@@ -1,3 +1,5 @@
+import {LinkMetdata} from "@lukso/lsp-factory.js";
+
 export interface Errors {
   search?: string
   amount?: string
@@ -12,6 +14,14 @@ export type NotificationType = 'primary' | 'danger' | 'info' | 'warning'
 
 export type Channel = 'browserExtension' | 'walletConnect'
 
+export enum ContractStandard {
+  LSP8 = 'LSP8',
+  LSP7 = 'LSP7',
+  ERC20 = 'ERC20',
+  ERC777 = 'ERC777',
+  ERC721 = 'ERC721',
+}
+
 export interface Store {
   isConnected: boolean
   address: string
@@ -19,4 +29,11 @@ export interface Store {
   balance: number
   channel?: Channel
   tokenAddress?: string
+}
+
+export type Lsp4Metadata = {
+  description: string
+  links: LinkMetdata[]
+  icon?: File
+  images?: File[]
 }
